int64

+++ internal/handlers/handler_test.go
@@ -210,7 +210,7 @@ func TestFind(t *testing.T) {
 				"sort", types.MustNewDocument(
 					"actor_id", int32(1),
 				),
-				"limit", int64(1),
+				"limit", int32(1),
 			),
 			resp: types.MustNewArray(
 				types.MustNewDocument(
@@ -368,7 +368,7 @@ func TestFind(t *testing.T) {
 				"sort", types.MustNewDocument(
 					"actor_id", int32(1),
 				),
-				"limit", int64(1),
+				"limit", int32(1),
 			),
 			resp: types.MustNewArray(
 				types.MustNewDocument(
